Rename files to match library rename
