+++ library.properties
@@ -1,9 +1,9 @@
-name=RTClibExtended
+name=OPEnS_RTC
 version=1.0.0
-author=FabioCuomo
-maintainer=FabioCuomo
-sentence=A fork of Adafruit RTC library and Jack Christensen's library
-paragraph=A fork of Adafruit RTC library and Jack Christensen's library
+author=OPEnS
+maintainer=OPEnS
+sentence=A fork of FabioCuomo-DS3231 and Adafruit RTC library and Jack Christensen's library
+paragraph=A fork of FabioCuomo-DS3231 and Adafruit RTC library and Jack Christensen's library
 category=Timing
-url=https://github.com/FabioCuomo/FabioCuomo-DS3231/
+url=https://github.com/OPEnSLab-OSU/OPEnS_RTC
 architectures=*
